Change the generic QueryNotExists() query to solution used for users
The problem is that the LDAP_sqlusers table and the gmail_users table
have different fields and the EXCEPT operator is failing.  Solution
involves executing the EXCEPT operator only on the pkey if both
tables.  This result set is then used as a subquery to determine who
needs to be updated.